# Perf improvement on Logger::Formatter#call